Updated photo importing

+++ app/sections/sql/log-notes.sql
@@ -7,9 +7,9 @@ WITH photo_notes AS (
 )
 SELECT
   id,
-  start_height,
-  end_height,
-  end_height-start_height span,
+  start_height::float,
+  end_height::float,
+  (end_height-start_height)::float span,
   coalesce(end_height > start_height, false) has_span,
   symbol,
   type,
